Social Network data
This data file contains 200 unique records, where each record is on its own line. For each record, there are 2 main components. The part before the colon is the name of the person for that particular record. The part after the colon is the names of that person's friends, whom they are in regular contact with. Each person has at least 1 friend, and each of the person's friends has also named that person as a friend. 

Each line of data in the file consists of the first and last name of a particular person, followed by a colon and a space character “: “ and then the first and last names of
each of their friends (people they are in regular contact with) with each friend’s name separated by a comma and a space “, ”.

Example of interpretation of data: The first line in the file is the record for Gill Bates. Gill Bates has named the following people as her social connections: Jodee Killam, Natacha Osterhoudt, Jom Tones, Verdie Tong, and Ossie Digangi. This means that if Gill Bates is contagious (able to spread the virus), Gill Bates may infect the people she has named, and if her friends are contagious, they may infect Gill Bates.